// apps/web/src/lib/api/generated/model/connectionInEnvironment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

/**
 * simulator: no network, every prompt is paid (not available in production)
 */
export type ConnectionInEnvironment = typeof ConnectionInEnvironment[keyof typeof ConnectionInEnvironment];


export const ConnectionInEnvironment = {
  sandbox: 'sandbox',
  production: 'production',
  simulator: 'simulator',
} as const;