// apps/web/src/lib/api/generated/model/connectionInShortcodeType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type ConnectionInShortcodeType = typeof ConnectionInShortcodeType[keyof typeof ConnectionInShortcodeType];


export const ConnectionInShortcodeType = {
  paybill: 'paybill',
  till: 'till',
} as const;